Separated the math module out and got it to compile properly
