+++ build.gradle
@@ -1,55 +1,49 @@
-plugins {
+plugins
+{
     id 'java'
-    id 'application'
-    
-  	id 'org.openjfx.javafxplugin' version '0.0.8'
-
-  	id 'eclipse'
+//    id 'application'
 }
 
-name = "Simulation-Q"
-group = "com.simulationQ"
 version = '1.3.0.0'
 
-repositories {
+repositories 
+{
     jcenter()
 }
 
-def currentOS = org.gradle.internal.os.OperatingSystem.current()
-def platform
-if (currentOS.isWindows()) {
-    platform = 'win'
-} else if (currentOS.isLinux()) {
-    platform = 'linux'
-} else if (currentOS.isMacOsX()) {
-    platform = 'mac'
-}
-dependencies {
-    implementation "org.openjfx:javafx-base:11:${platform}"
-    implementation "org.openjfx:javafx-graphics:11:${platform}"
-    implementation "org.openjfx:javafx-controls:11:${platform}"
-
-  testImplementation 'org.junit.jupiter:junit-jupiter-api:5.3.1'
-  testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.3.1'
-
+dependencies 
+{
+    implementation project( ':math' )
+    
+    testImplementation ('org.junit.jupiter:junit-jupiter-api:5.7.0')
+    testRuntimeOnly ('org.junit.jupiter:junit-jupiter-engine:5.7.0')
 }
 
-test{
-  useJUnitPlatform()
+test
+{
+    useJUnitPlatform()
 }
 
-javafx {
-    version = "11"
-    modules = [ 'javafx.controls', 'javafx.fxml' ]
-}
-
-run {
-//    if (javafx.platform.classifier == 'win') {
-//        // Temporal fix for Eclipse with JDK 1.8 and Windows
-//        systemProperty "java.library.path", "C:\tmp"
-//    }
-}
-
-application {
-	mainClassName 'com.Main'
+subprojects
+{
+    apply plugin: 'java'
+    apply plugin: 'eclipse'
+    
+    group = "com.simulationQ"
+    
+    repositories 
+    {
+        jcenter()
+    }
+    
+    dependencies
+    {
+        testImplementation ('org.junit.jupiter:junit-jupiter-api:5.7.0')
+        testRuntimeOnly ('org.junit.jupiter:junit-jupiter-engine:5.7.0')
+    }
+    
+    test
+    {
+        useJUnitPlatform()
+    }
 }
