+++ math/.gitignore
@@ -1,0 +1,13 @@
+/bin
+/test_results
+/.settings
+
+.project
+.classpath
+
+# Ignore Gradle project-specific cache directory
+.gradle
+gradle.properties
+
+# Ignore Gradle build output directory
+/build
