--- src/main/java/com/gui/events/package-info.java
@@ -1,8 +1,0 @@
-/*
- * 22/03/2020 13:51:00
- * package-info.java created by Tsvetelin
- */
-/**
- * @author Tsvetelin
- */
-package com.gui.events;
