--- src/main/java/com/gui/package-info.java
@@ -1,8 +1,0 @@
-/*
- * 21/03/2020 14:43:44
- * package-info.java created by Tsvetelin
- */
-/**
- * @author Tsvetelin
- */
-package com.gui;
